Printing distance travelled in millimeters on LCD
The getDistanceTravelled function asks for data from the robot so when
no robot is connected it freezes the program, therefore it is commented
for now. Uncomment if connectiong to robot.

The setupLCD changes the port A to analouge and that might FK things up
with the robot, don't know yet. If it does just comment the setupLCD for
now.